requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
bcrypt==4.0.1
email-validator==2.1.0
requests==2.31.0
PyJWT==2.8.0
redis==5.0.1
celery==5.3.4